Remove most use of uint32_t, to avoid programming errors

+++ src/gtil/output_shape.cc
@@ -17,20 +17,21 @@ namespace treelite::gtil {
 std::vector<std::uint64_t> GetOutputShape(
     Model const& model, std::uint64_t num_row, Configuration const& config) {
   auto const num_tree = model.GetNumTree();
-  auto const max_num_class
-      = *std::max_element(model.num_class.Data(), model.num_class.Data() + model.num_target);
+  auto const max_num_class = static_cast<std::uint64_t>(
+      *std::max_element(model.num_class.Data(), model.num_class.Data() + model.num_target));
   switch (config.pred_type) {
   case PredictKind::kPredictDefault:
   case PredictKind::kPredictRaw:
     if (model.num_target > 1) {
-      return {model.num_target, num_row, max_num_class};
+      return {static_cast<std::uint64_t>(model.num_target), num_row, max_num_class};
     } else {
       return {num_row, max_num_class};
     }
   case PredictKind::kPredictLeafID:
     return {num_row, num_tree};
   case PredictKind::kPredictPerTree:
-    return {num_row, num_tree, model.leaf_vector_shape[0] * model.leaf_vector_shape[1]};
+    return {num_row, num_tree,
+        static_cast<std::uint64_t>(model.leaf_vector_shape[0]) * model.leaf_vector_shape[1]};
   default:
     TREELITE_LOG(FATAL) << "Unsupported model type: " << static_cast<int>(config.pred_type);
     return {};
